requirements.txt
streamlit==1.38.0
tensorflow-cpu==2.16.0
numpy==2.1.1
opencv-python-headless==4.8.0.76
Pillow==10.4.0
